+++ jiyoontrip/src/app/components/mypage/password-change/styles.module.css
@@ -9,10 +9,11 @@
   display: flex;
   width: 1280px;
   height: 24px;
-  font-family: "Pretendard Variable";
+  font-family: "Pretendard Variable", sans-serif;
   font-weight: 600;
   font-size: 18px;
   line-height: 24px;
+  letter-spacing: 0;
   color: #000000;
 }
 
@@ -47,18 +48,20 @@
 }
 
 .label {
-  font-family: "Pretendard Variable";
+  font-family: "Pretendard Variable", sans-serif;
   font-weight: 500;
   font-size: 16px;
   line-height: 24px;
+  letter-spacing: 0;
   color: #333333;
 }
 
 .required {
-  font-family: "Pretendard Variable";
+  font-family: "Pretendard Variable", sans-serif;
   font-weight: 500;
   font-size: 16px;
   line-height: 24px;
+  letter-spacing: 0;
   color: #f66a6a;
   margin-left: 4px;
 }
